refactor(config): document import-surface move; rename cache_lib->cache_stats
双轨 CR re-review follow-up on PR #162:

- Note the import-surface break for the published package in both CHANGELOGs
  ([Unreleased] -> Changed): activation/cache symbols no longer import from
  config; tool surface unchanged. A re-export shim is intentionally NOT added
  (it would recreate the activation<->config cycle).
- Rename cache_lib.py -> cache_stats.py (mirrors the TS cacheStats.ts naming;
  STYLE discourages the generic -lib suffix) and repoint the 8 importers.
- Expand the TS cycle-invariant comment with the rejection reasoning for the
  Bot's "extract state" suggestion: the cycle is between the LOGIC
  (config.loadConfig <-> activation.checkActivationChange/withActivationSnapshot),
  not the state -- relocating mutable vars would not remove those call edges
  without loadConfig dropping its auto-check (a behavior change). Hence the
  invariant comment, not a structural split (the Bot's accepted fallback).

Behavior-preserving: Python 369 / TS 283 green, typecheck clean.
