+++ python/src/prts_mcp/data/stage.py
@@ -5,7 +5,7 @@ from dataclasses import dataclass as _dataclass
 from typing import Any
 
 from prts_mcp.activation import register_activation_listener
-from prts_mcp.cache_lib import (
+from prts_mcp.cache_stats import (
     activation_aware_cache as _activation_aware_cache,
     cache_stat as _cache_stat,
 )
