+++ python/tests/test_config.py
@@ -8,7 +8,7 @@ from unittest.mock import patch
 
 import prts_mcp.config as config_module
 from prts_mcp.activation import activation_snapshot
-from prts_mcp.cache_lib import activation_aware_cache
+from prts_mcp.cache_stats import activation_aware_cache
 from prts_mcp.config import Config
